price_manager/requirements.txt
pandas>=1.3.0
django>=5.2.4
django-tables2>=2.7.5
openpyxl>=3.1.5
xlrd>=2.0.1
psycopg2>=2.9.10
tablib>=3.8.0
django-import-export>=4.3.9
django-filter>=25.1
django-autocomplete-light>=3.12.1
django-htmx>=1.27.0
django-template-partials>=25.3